+++ examples/AN00121_udp_demo/src/test_udp.py
@@ -1,4 +1,6 @@
 #!/usr/bin/python
+# Copyright 2015-2025 XMOS LIMITED.
+# This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 import argparse
 import socket