+++ examples/AN00121_udp_demo/src/udp_reflect.h
@@ -1,4 +1,5 @@
-// Copyright (c) 2011-2016, XMOS Ltd, All rights reserved
+// Copyright 2011-2025 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 #ifndef __UDP_REFLECT_H__
 #define __UDP_REFLECT_H__