+++ lib_xtcp/api/xtcp.h
@@ -1,4 +1,5 @@
-// Copyright (c) 2011-2017, XMOS Ltd, All rights reserved
+// Copyright 2011-2025 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
 #ifndef __xtcp_h__
 #define __xtcp_h__