+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 TCP/IP library change log
 =========================
 
+UNRELEASED
+----------
+
+  * ADDED: CMake support
+  * FIXED: Many compiler warnings.
+  * CHANGE: Moved app note examples out to own repos, AN00121 
+    (XK_ETH_XU316_DUAL_100M) and AN02044 (XCORE-200-EXPLORER)
+  * CHANGE: Removed 'smi_if i_smi' and 'phy_address' parameters from both
+    xtcp_uip and xtcp_lwip functions. The smi_if now only links the smi 
+    component with the PHY driver.
+
 6.1.0
 -----